chore: rename the slice to a more clear name
