+++ packages/compass-indexes/src/modules/index-view.ts
